Optimized preemption control code (#81)
This change migrates the existing preemption control code implementation
to utilize TCTs for wait for AIE DMA completions. Also, enable all the
available DMA channels to utilize full DMA bandwidth.

+++ test/cpp_test/c_api/aie_test_common.h
@@ -7,7 +7,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-char* ReadFile(char *name, long *s)
+char* ReadFile(char *name, size_t *s)
 {
   FILE *file;
   char *buffer;
